# openapi.json
{
  "components": {
    "schemas": {
...
      "AdapterDescription": {
        "properties": {
          "capabilities": true,
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parameters": true,
          "releases": {
            "description": "Official Linux builds offered in the flight sheet, newest first. Empty for adapters\nthat only run a custom download URL.",
            "items": {
              "$ref": "#/components/schemas/MinerRelease"
            },
            "type": "array"
          },
          "schema_version": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "schema_version",
          "parameters",
          "capabilities",
          "releases"
        ],
        "title": "AdapterDescription",
        "type": "object"
      },
...
      "AuditEvent": {
        "properties": {
          "action": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "detail": true,
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "target": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "actor",
          "action",
          "target",
          "detail",
          "created_at"
        ],
        "title": "AuditEvent",
        "type": "object"
      },
...
      "CatalogInput": {
        "properties": {
          "data": true,
          "expected_revision": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "data"
        ],
        "title": "CatalogInput",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "session": {
        "in": "cookie",
        "name": "rigdeck_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "RigDeck API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/adapters": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdapterDescription"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden or invalid CSRF"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/audit": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEvent"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden or invalid CSRF"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/catalog/{kind}": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogItem"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden or invalid CSRF"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "kind",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden or invalid CSRF"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/catalog/{kind}/{id}": {
      "delete": {
        "responses": {
          "204": {
            "description": "Deleted"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "kind",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden or invalid CSRF"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/flight-sheets": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FlightSheet"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden or invalid CSRF"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightSheet"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden or invalid CSRF"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/flight-sheets/{id}": {
      "delete": {
        "responses": {
          "204": {
            "description": "Deleted"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightSheet"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden or invalid CSRF"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
...
  }
}